refactor: faceSquares and FRCamera components moved in different files from FaceRecognition screen
